+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [fix: fix lb configuration on bootstrap step.](https://github.com/Tencent/spring-cloud-tencent/issues/1706)
 - [feat:support fault injection.](https://github.com/Tencent/spring-cloud-tencent/pull/1707)
 - [feat: support config event and monitor address list.](https://github.com/Tencent/spring-cloud-tencent/pull/1708)
+- [fix:replace HttpClient with HttpURLConnection for JDK 8 compatibility.](https://github.com/Tencent/spring-cloud-tencent/pull/1709)